Eliminate p6scalarfromdesc ext-op for a desugar
No semantic change here, just a change in how we achieve it. This is
both potentially slower (sequence of VM instructions rather than a C
op) but potentially faster (spesh can eliminate a branch that it could
not when it was just calling a chunk of opaque C code).

This does also make the slight improvement that we only ever create one
default container descriptor now, whereas before we might have had many
copies of it serialized across different precompilations.